+++ src/test/java/racingcar/IntScannerTest.java
@@ -4,6 +4,7 @@ import java.io.ByteArrayInputStream;
 import java.io.InputStream;
 import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.jupiter.api.Test;
+import racingcar.module.IntScanner;
 
 public class IntScannerTest {
     void setInputSteam(String input) {
